requirements.txt
grpcio==1.65.5
grpcio-tools==1.65.5
protobuf==5.27.3
grpc-reflection==1.0.0
